requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.9.3
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.1.0
frozenlist==1.3.3
googlesearch-python==1.1.0
idna==2.10
multidict==6.0.4
requests==2.25.1
soupsieve==2.4
urllib3==1.26.15
yarl==1.8.2
